requirements.txt
torch>=1.1.0
torchvision>=0.3.0
tqdm==4.36.1
opencv-python==3.4.3.18
Pillow>=6.2.0
PrettyTable>=0.7.2
matplotlib>=3.1.1
tensorboardX>=1.9
